backend/requirements.txt
# 后端 Python 依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
redis==5.0.1
cryptography==41.0.8

# AI 和机器学习
langchain==0.0.350
langchain-community==0.0.7
openai==1.3.7
anthropic==0.7.8
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2

# 向量数据库
chromadb==0.4.18
qdrant-client==1.7.0
faiss-cpu==1.7.4
pinecone-client==2.2.4
weaviate-client==3.25.3

# 多模态处理
pillow==10.1.0
opencv-python==4.8.1.78
librosa==0.10.1
python-multipart==0.0.6
python-magic==0.4.27

# 文档处理
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2
chardet==5.2.0

# 网络和通信
aiohttp==3.9.1
websockets==12.0
httpx==0.25.2

# 工具和实用程序
celery==5.3.4
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 开发和测试工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# IM 平台适配
requests==2.31.0
qrcode==7.4.2